Clean up examples to remove warnings

+++ 8.Strings/StringLength/StringLength.ino
@@ -12,7 +12,7 @@
  This example code is in the public domain.
  */
 String txtMsg = "";                         // a string for incoming text
-int lastStringLength = txtMsg.length();     // previous length of the String
+size_t lastStringLength = txtMsg.length();     // previous length of the String
 
 void setup() {
   // open the serial port:
